Improved Edge Processing
The cuda code now detects a radius around the curb cells and retains the point cloud density in those cells as well. This allows the edges to be retained better in the quadric decimation process in vr mesh. The C code does not do this yet, that will be a future edit